+++ back/src/main/java/com/ll/goohaeyou/member/member/domain/MemberDomainService.java
@@ -1,0 +1,67 @@
+package com.ll.goohaeyou.member.member.domain;
+
+import com.ll.goohaeyou.auth.exception.AuthException;
+import com.ll.goohaeyou.global.standard.anotations.DomainService;
+import com.ll.goohaeyou.global.standard.base.util.Util;
+import com.ll.goohaeyou.member.member.application.dto.JoinRequest;
+import com.ll.goohaeyou.member.member.domain.entity.Member;
+import com.ll.goohaeyou.member.member.domain.repository.MemberRepository;
+import com.ll.goohaeyou.member.member.domain.type.Role;
+import com.ll.goohaeyou.member.member.exception.MemberException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.transaction.annotation.Transactional;
+
+@DomainService
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MemberDomainService {
+    private final MemberRepository memberRepository;
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    @Transactional
+    public void joinMember(JoinRequest request) {
+        memberRepository.findByUsername(request.username())
+                .ifPresent(member -> {
+                    throw new MemberException.DuplicateUsernameException();
+                });
+
+        Role role = "admin".equals(request.username()) ? Role.ADMIN : Role.USER;
+
+        if (role == Role.ADMIN) {
+            memberRepository.save(
+                    Member.createAdmin(
+                            bCryptPasswordEncoder.encode(request.password()),
+                            request.name(),
+                            request.phoneNumber(),
+                            request.email(),
+                            request.gender(),
+                            request.location(),
+                            request.birth(),
+                            Util.Region.getRegionCodeFromAddress(request.location())
+                    )
+            );
+        }
+        if (role == Role.USER) {
+            memberRepository.save(
+                    Member.createUser(
+                            request.username(),
+                            bCryptPasswordEncoder.encode(request.password()),
+                            request.name(),
+                            request.phoneNumber(),
+                            request.email(),
+                            request.gender(),
+                            request.location(),
+                            request.birth(),
+                            Util.Region.getRegionCodeFromAddress(request.location())
+                    )
+            );
+        }
+    }
+
+    public void verifyPassword(String MemberPassword, String inputPassword) {
+        if (!bCryptPasswordEncoder.matches(MemberPassword, inputPassword)) {
+            throw new AuthException.InvalidPasswordException();
+        }
+    }
+}
