+++ back/src/main/java/com/ll/goohaeyou/member/member/domain/MemberReader.java
@@ -1,0 +1,20 @@
+package com.ll.goohaeyou.member.member.domain;
+
+import com.ll.goohaeyou.global.standard.anotations.DomainService;
+import com.ll.goohaeyou.member.member.domain.entity.Member;
+import com.ll.goohaeyou.member.member.domain.repository.MemberRepository;
+import com.ll.goohaeyou.member.member.exception.MemberException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
+
+@DomainService
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MemberReader {
+    private final MemberRepository memberRepository;
+
+    public Member getMemberByUsername(String username) {
+        return memberRepository.findByUsername(username)
+                .orElseThrow(MemberException.MemberNotFoundException::new);
+    }
+}
